BufferingBlockCreator -> NetworkReceiver.BlockGenerator
